third_party/blink/web_tests/platform/generic/svg/animations/animate-endElement-beginElement-expected.txt

SVG 1.1 dynamic animation tests

Tests animation beginElement command's restarting capability after endElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rect.x.animVal.value is 50
PASS rect.x.baseVal.value is 0
PASS successfullyParsed is true

TEST COMPLETE
